default first item to 90 degrees front and center
This is is still not ideal, but it is better if there are only one or two items
in the launcher.  This way, they will show up and not be somewhere off the
screen initially.  The improved approach would be to center all available items
at 90 degrees, and then secondly solve the wrap issue where everything is moved
off the screen (or just disallow scrolling).

Signed-off-by: Joshua Henderson <joshua.henderson@microchip.com>